Parse table names
Test data tables are recognized from one or more asterisks (*), followed by a normal table name and an optional closing asterisks.
The test data tables must have one or more asterisk before their names similarly as in the TSV format.
All tables that do not start with a recognized table name should be ignored.
Otherwise asterisks and possible spaces in the table header are ignored so, for example, *** Settings *** and *Settings work the same way.